requirements.txt
absl_py
cloudpickle
fairscale
faiss_gpu
json_repair
peft
pydantic
pydantic_core
PyYAML
sentence_transformers
sentencepiece
setproctitle
tensorboardX
numpy==1.26.3
xformers==0.0.27
# for the torch version, below is the one I use.
# you can choose any other versions that work.
https://download.pytorch.org/whl/cu121/torch-2.3.1%2Bcu121-cp310-cp310-linux_x86_64.whl#sha256=f0deb5d2f932a68ed54625ba140eddbf2af22be978ee19b9b63c986add6425b2
https://download.pytorch.org/whl/cu121/torchvision-0.18.1%2Bcu121-cp310-cp310-linux_x86_64.whl#sha256=e95ba5a2c616939281e01babf11664d6d1725e81bba57ef81f81c3e57e4d4151
https://download.pytorch.org/whl/cu121/torchaudio-2.3.1%2Bcu121-cp310-cp310-linux_x86_64.whl#sha256=0b423f4ae3356f11f6723e8c77208ac3f9361a4f941e4cc08d86c32c137594bc